move `shareApp` method from settings screen
